display dataset list only if the user is connected
