+++ client/src/constants.ts
@@ -1,6 +1,8 @@
 import type { DataFormat, UpdateFrequency } from "./definitions/datasets";
 
-export const PUBLIC_PAGES = ["/login"];
+
+// Note : the / page (the home page) is "semi public". This page layout and information displayed depends whether the user is connected or not
+export const PUBLIC_PAGES = ["/login", "/"];
 
 export const DATA_FORMAT_LABELS: { [K in DataFormat]: string } = {
   file_tabular: "Fichier tabulaire (XLS, XLSX, CSV, ...)",
